rename: 가입 대기 회원 관련 컴포넌트 PendingMembers 폴더로 이동
