+++ src/pages/PendingMembersPage.tsx
@@ -1,5 +1,5 @@
-import PendingMembersHeader from "@/components/common/Header/PendingMembersHeader";
-import PendingMembersInfoTable from "@/components/common/Table/PendingMembersInfoTable";
+import PendingMembersHeader from "@/components/PendingMembers/PendingMembersHeader";
+import PendingMembersInfoTable from "@/components/PendingMembers/PendingMembersInfoTable";
 import Title from "@/components/common/Title";
 import PendingMembersContextProvider from "@/contexts/PendingMembersContext";
